IGNITE-24234 Migration-Tools: Move tests requiring docker to the integration phase
