+++ migration-tools/modules/migration-tools-config-converter/build.gradle
@@ -18,6 +18,8 @@
 apply from: "$rootDir/buildscripts/java-core.gradle"
 apply from: "$rootDir/buildscripts/publishing.gradle"
 apply from: "$rootDir/buildscripts/java-junit5.gradle"
+apply from: "$rootDir/buildscripts/java-integration-test.gradle"
+apply from: "$rootDir/buildscripts/java-test-fixtures.gradle"
 
 description = 'migration-tools-config-converter'
 
@@ -31,12 +33,22 @@ dependencies {
     implementation libs.mockito.core
     implementation libs.slf4j.api
 
+    testImplementation testFixtures(project(":migration-tools-config-converter"))
     testImplementation project(path: ":migration-tools-ignite3-repack", configuration: "shadow")
     testImplementation project(":migration-tools-commons-tests")
     testImplementation libs.assertj.core
     testImplementation libs.junit5.api
     testImplementation libs.junit5.params
-    testImplementation libs.testcontainers.core
-    testImplementation libs.testcontainers.junit
     testImplementation libs.typesafe.config
+
+    integrationTestImplementation testFixtures(project(":migration-tools-config-converter"))
+    integrationTestImplementation project(":migration-tools-commons-tests")
+    integrationTestImplementation libs.assertj.core
+    integrationTestImplementation libs.junit5.api
+    integrationTestImplementation libs.junit5.params
+    integrationTestImplementation libs.testcontainers.core
+    integrationTestImplementation libs.testcontainers.junit
+
+    testFixturesImplementation libs.commons.lang3
+    testFixturesImplementation libs.typesafe.config
 }
