+++ migration-tools/modules/migration-tools-persistence/build.gradle
@@ -18,6 +18,7 @@
 apply from: "$rootDir/buildscripts/java-core.gradle"
 apply from: "$rootDir/buildscripts/publishing.gradle"
 apply from: "$rootDir/buildscripts/java-junit5.gradle"
+apply from: "$rootDir/buildscripts/java-integration-test.gradle"
 
 description = 'migration-tools-persistence'
 
@@ -40,8 +41,23 @@ dependencies {
     testImplementation libs.assertj.core
     testImplementation libs.junit5.api
     testImplementation libs.junit5.params
-    testImplementation libs.testcontainers.core
-    testImplementation libs.testcontainers.junit
+
+    integrationTestImplementation project(":migration-tools-commons")
+    integrationTestImplementation project(path: ":migration-tools-ignite3-repack", configuration: "shadow")
+    integrationTestImplementation project(":migration-tools-commons-tests")
+    integrationTestImplementation project(":migration-tools-e2e-core")
+    integrationTestImplementation project(":migration-tools-e2e-implementations")
+    integrationTestImplementation testFixtures(project(":migration-tools-ai2-test-fixtures"))
+    integrationTestImplementation libs.ignite2.core
+    integrationTestImplementation libs.commons.lang3
+    integrationTestImplementation libs.commons.collections4
+    integrationTestImplementation libs.jackson.databind
+    integrationTestImplementation libs.mockito.core
+    integrationTestImplementation libs.assertj.core
+    integrationTestImplementation libs.junit5.api
+    integrationTestImplementation libs.junit5.params
+    integrationTestImplementation libs.testcontainers.core
+    integrationTestImplementation libs.testcontainers.junit
 }
 
-tasks.test.dependsOn += ":migration-tools-e2e-ai2-runner:docker"
+tasks.integrationTest.dependsOn += ":migration-tools-e2e-ai2-runner:docker"
