+++ migration-tools/modules/migration-tools-persistence/src/integrationTest/java/org/apache/ignite/migrationtools/persistence/CacheDdlCreationFromWorkDirTest.java
