+++ migration-tools/modules/migration-tools-persistence/src/test/java/org/apache/ignite/migrationtools/persistence/SchemaColumnsProcessorTest.java
@@ -19,7 +19,7 @@ package org.apache.ignite.migrationtools.persistence;
 
 import static org.apache.ignite.migrationtools.sql.SqlDdlGenerator.EXTRA_FIELDS_COLUMN_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.ArrayList;
 import java.util.Collections;
